ai_engine/requirements.txt
opencv-python
mediapipe
mariadb
Flask
python-dotenv
